Convert ITAggregateQueryTest to a unit test
